// frontend/src/generated/models/Account.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Account = {
    account_type: string;
    available_balance?: number | null;
    balance: number;
    created_at: string;
    id: string;
    institution: string;
    is_credit_card?: boolean | null;
    last_updated: string;
    name: string;
    simplefin_id?: string | null;
};